+++ packages/expo-prevent-screen-capture/android/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
-<manifest package="expo.modules.preventscreencapture">
+<manifest package="expo.modules.screencapture">
 </manifest>
