+++ packages/expo-prevent-screen-capture/android/src/main/kotlin/expo/modules/screencapture/ScreenCapturePackage.kt
@@ -1,4 +1,4 @@
-package expo.modules.preventscreencapture
+package expo.modules.screencapture
 
 import android.content.Context
 
@@ -6,9 +6,9 @@ import org.unimodules.core.BasePackage
 import org.unimodules.core.ExportedModule
 import org.unimodules.core.ViewManager
 
-class PreventScreenCapturePackage : BasePackage() {
+class ScreenCapturePackage : BasePackage() {
   override fun createExportedModules(context: Context): List<ExportedModule> {
-    return listOf(PreventScreenCaptureModule(context) as ExportedModule)
+    return listOf(ScreenCaptureModule(context) as ExportedModule)
   }
 }
