--- packages/expo-prevent-screen-capture/src/ExpoPreventScreenCapture.ts
@@ -1,3 +1,0 @@
-import { NativeModulesProxy } from '@unimodules/core';
-
-export default NativeModulesProxy.ExpoPreventScreenCapture;
