+++ packages/expo-prevent-screen-capture/src/ExpoScreenCapture.ts
@@ -1,0 +1,3 @@
+import { NativeModulesProxy } from '@unimodules/core';
+
+export default NativeModulesProxy.ExpoScreenCapture;
