+++ packages/expo-prevent-screen-capture/src/__tests__/ScreenCapture-test.js
@@ -1,0 +1,15 @@
+import * as ScreenCapture from '../ScreenCapture';
+
+describe('ScreenCapture methods are defined', () => {
+  it('activatePreventScreenCapture is defined', async () => {
+    expect(ScreenCapture.activatePreventScreenCapture).toBeDefined();
+  });
+
+  it('deactivatePreventScreenCapture is defined', async () => {
+    expect(ScreenCapture.deactivatePreventScreenCapture).toBeDefined();
+  });
+
+  it('usePreventScreenCapture hook is defined', async () => {
+    expect(ScreenCapture.usePreventScreenCapture).toBeDefined();
+  });
+});
