+++ packages/expo-prevent-screen-capture/unimodule.json
@@ -1,4 +1,4 @@
 {
-  "name": "expo-prevent-screen-capture",
+  "name": "expo-screen-capture",
   "platforms": ["ios", "android", "web"]
 }
